pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "vultronscanner"
version = "0.1.0"
description = "Modular Attack Surface Manager & Penetration Testing Platform"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = [
    "penetration-testing",
    "network-security",
    "vulnerability-scanner",
    "attack-surface",
    "cybersecurity",
    "nmap",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
]
dependencies = [
    "aiohttp>=3.9.5",
    "aiofiles>=23.2.1",
    "httpx>=0.27.0",
    "python-nmap>=0.7.1",
    "dnspython>=2.6.1",
    "ipwhois>=1.3.0",
    "requests>=2.31.0",
    "pydantic>=2.7.1",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
    "click>=8.1.7",
    "jinja2>=3.1.4",
    "cvss>=3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "black>=24.3.0",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "mypy>=1.9.0",
    "bandit>=1.7.8",
    "pre-commit>=3.7.0",
]

[project.scripts]
vultron = "main:cli"

[project.urls]
Homepage = "https://github.com/iremysv/VultronScanner"
Repository = "https://github.com/iremysv/VultronScanner"
Issues = "https://github.com/iremysv/VultronScanner/issues"
Changelog = "https://github.com/iremysv/VultronScanner/blob/main/CHANGELOG.md"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["Tests"]
addopts = "-v"

[tool.coverage.run]
omit = ["Tests/*", ".venv/*", "setup.py"]
